use extend api

+++ src/main/java/re/imc/xreplayminio/XReplayMinio.java
@@ -1,16 +1,18 @@
 package re.imc.xreplayminio;
 
-import de.musterbukkit.replaysystem.main.ReplayAPI;
 import io.minio.MinioClient;
 import org.bukkit.Bukkit;
 import org.bukkit.configuration.ConfigurationSection;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.Listener;
+import org.bukkit.event.server.PluginEnableEvent;
 import org.bukkit.plugin.java.JavaPlugin;
 import re.imc.xreplayminio.command.ReplayLoadCommand;
 import re.imc.xreplayminio.listener.ReplayListener;
 
 import java.io.File;
 
-public final class XReplayMinio extends JavaPlugin {
+public final class XReplayMinio extends JavaPlugin implements Listener {
 
     private static XReplayMinio instance;
 
@@ -22,13 +24,21 @@ public final class XReplayMinio extends JavaPlugin {
         instance = this;
 
         loadMinio();
-        if (getConfig().getBoolean("use-default-world", false)) {
-            ReplayAPI.setMapName(Bukkit.getWorlds().get(0).getName());
+
+        Bukkit.getPluginManager().registerEvents(this, this);
+    }
+
+    @EventHandler
+    public void onReplayPluginLoad(PluginEnableEvent event) {
+
+        if (!event.getPlugin().getName().equalsIgnoreCase("replaysystem")) {
+            return;
         }
+
         Bukkit.getPluginManager().registerEvents(new ReplayListener(), this);
         Bukkit.getPluginCommand("replayload").setExecutor(new ReplayLoadCommand());
         deleteDir(new File(XReplayMinio.getInstance().getConfig()
-                    .getString("target-folder")));
+                .getString("target-folder")));
     }
 
     public static void deleteDir(File file) {
@@ -47,18 +57,16 @@ public final class XReplayMinio extends JavaPlugin {
 
     }
 
+
     @Override
     public void onDisable() {
         // Plugin shutdown logic
-        ReplayAPI.saveReplay();
-        try {
-            Thread.sleep(1000 * 6);
-        } catch (InterruptedException e) {
-            throw new RuntimeException(e);
-        }
+        System.out.println("WHAT");
         File folder = new File(XReplayMinio.getInstance().getConfig()
                 .getString("target-folder"));
         File[] files = folder.listFiles();
+
+        System.out.println(files.length);
         if (files != null && files.length > 0) {
             getLogger().info("Finally Replay");
             while (true) {
@@ -67,16 +75,13 @@ public final class XReplayMinio extends JavaPlugin {
                 } catch (InterruptedException e) {
                     throw new RuntimeException(e);
                 }
-                try {
-                    Thread.sleep(1000);
-                } catch (InterruptedException e) {
-                    throw new RuntimeException(e);
-                }
 
                 XReplayMinio.getInstance().getLogger().info("Compressing...");
             }
         }
     }
+
+
 
     public static XReplayMinio getInstance() {
         return instance;
